EDITOR: Reorganization and minor efficiency improvements (#886)
