+++ package.json
@@ -10,30 +10,30 @@
     "url": "https://github.com/bitburner-official/bitburner-src/issues"
   },
   "dependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.10.5",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
     "@material-ui/core": "^4.12.4",
-    "@mui/icons-material": "^5.11.0",
+    "@mui/icons-material": "^5.14.12",
     "@mui/material": "^5.14.12",
-    "@mui/styles": "^5.11.2",
-    "@mui/system": "^5.0.3",
-    "@types/estree": "^1.0.0",
-    "@types/react-syntax-highlighter": "^15.5.7",
-    "acorn": "^8.8.1",
+    "@mui/styles": "^5.14.12",
+    "@mui/system": "^5.14.12",
+    "@types/estree": "^1.0.2",
+    "@types/react-syntax-highlighter": "^15.5.8",
+    "acorn": "^8.10.0",
     "acorn-walk": "^8.2.0",
     "arg": "^5.0.2",
     "bcryptjs": "^2.4.3",
-    "better-react-mathjax": "^2.0.2",
+    "better-react-mathjax": "^2.0.3",
     "buffer": "^6.0.3",
     "clsx": "^1.2.1",
-    "date-fns": "^2.29.3",
-    "escodegen": "^2.0.0",
+    "date-fns": "^2.30.0",
+    "escodegen": "^2.1.0",
     "file-saver": "^2.0.5",
-    "jquery": "^3.6.3",
+    "jquery": "^3.7.1",
     "js-sha256": "^0.9.0",
     "jszip": "^3.10.1",
     "material-ui-color": "^1.2.0",
-    "material-ui-popup-state": "^1.5.3",
+    "material-ui-popup-state": "^1.9.3",
     "monaco-vim": "^0.3.5",
     "notistack": "^2.0.8",
     "numeral": "^2.0.6",
@@ -41,63 +41,63 @@
     "react": "^17.0.2",
     "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^17.0.2",
-    "react-draggable": "^4.4.4",
+    "react-draggable": "^4.4.6",
     "react-markdown": "^8.0.7",
-    "react-resizable": "^3.0.4",
+    "react-resizable": "^3.0.5",
     "react-syntax-highlighter": "^15.5.0",
     "remark-gfm": "^3.0.1",
-    "sprintf-js": "^1.1.2"
+    "sprintf-js": "^1.1.3"
   },
   "description": "A cyberpunk-themed incremental game",
   "devDependencies": {
-    "@babel/core": "^7.20.12",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.18.6",
-    "@microsoft/api-documenter": "^7.21.3",
-    "@microsoft/api-extractor": "^7.34.2",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
-    "@types/bcryptjs": "^2.4.2",
+    "@babel/core": "^7.23.0",
+    "@babel/preset-env": "^7.22.20",
+    "@babel/preset-react": "^7.22.15",
+    "@babel/preset-typescript": "^7.23.0",
+    "@microsoft/api-documenter": "^7.23.9",
+    "@microsoft/api-extractor": "^7.38.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.11",
+    "@types/bcryptjs": "^2.4.4",
     "@types/escodegen": "^0.0.7",
     "@types/file-saver": "^2.0.5",
-    "@types/jest": "^29.4.0",
-    "@types/jquery": "^3.5.16",
-    "@types/lodash": "^4.14.191",
-    "@types/numeral": "^2.0.2",
-    "@types/react": "^17.0.52",
-    "@types/react-beautiful-dnd": "^13.1.3",
-    "@types/react-dom": "^17.0.20",
-    "@types/react-resizable": "^3.0.3",
-    "@typescript-eslint/eslint-plugin": "^5.48.0",
-    "@typescript-eslint/parser": "^5.48.0",
-    "babel-jest": "^29.3.1",
-    "babel-loader": "^9.1.2",
-    "css-loader": "^6.7.3",
-    "electron": "^26.2.4",
-    "electron-packager": "^17.1.1",
+    "@types/jest": "^29.5.5",
+    "@types/jquery": "^3.5.22",
+    "@types/lodash": "^4.14.199",
+    "@types/numeral": "^2.0.3",
+    "@types/react": "^17.0.67",
+    "@types/react-beautiful-dnd": "^13.1.5",
+    "@types/react-dom": "^17.0.21",
+    "@types/react-resizable": "^3.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "babel-jest": "^29.7.0",
+    "babel-loader": "^9.1.3",
+    "css-loader": "^6.8.1",
+    "electron": "^26.3.0",
+    "electron-packager": "^17.1.2",
     "eslint": "^8.51.0",
-    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "file-loader": "^6.2.0",
-    "fork-ts-checker-webpack-plugin": "^7.2.14",
-    "html-webpack-plugin": "^5.5.0",
+    "fork-ts-checker-webpack-plugin": "^7.3.0",
+    "html-webpack-plugin": "^5.5.3",
     "http-server": "^14.1.1",
-    "jest": "^29.4.3",
-    "jest-environment-jsdom": "^29.4.3",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "jsdom": "^20.0.3",
     "lodash": "^4.17.21",
     "monaco-editor": "^0.35.0",
-    "monaco-editor-webpack-plugin": "^7.0.1",
-    "prettier": "^2.8.1",
+    "monaco-editor-webpack-plugin": "^7.1.0",
+    "prettier": "^2.8.8",
     "raw-loader": "^4.0.2",
     "react-refresh": "^0.14.0",
     "source-map": "^0.7.4",
-    "start-server-and-test": "^1.15.2",
-    "style-loader": "^3.3.1",
+    "start-server-and-test": "^1.15.4",
+    "style-loader": "^3.3.3",
     "typescript": "^5.2.2",
-    "webpack": "^5.75.0",
-    "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.11.1"
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   },
   "engines": {
     "node": ">=14"
